Move eff.py to _internal
