+++ certbot/_internal/client.py
@@ -23,7 +23,7 @@ from certbot._internal import auth_handler
 from certbot import cli
 from certbot import constants
 from certbot import crypto_util
-from certbot import eff
+from certbot._internal import eff
 from certbot._internal import error_handler
 from certbot import errors
 from certbot import interfaces
