--- docs/api/eff.rst
@@ -1,5 +1,0 @@
-:mod:`certbot.eff`
-----------------------
-
-.. automodule:: certbot.eff
-   :members:
